model update: sensor readings are decimal now

+++ server_side_v2/PlantMonitoringSystem/PlantMonitoringSystem.Database/sensorreading.cs
@@ -15,7 +15,7 @@ namespace PlantMonitoringSystem.Database
     public partial class sensorreading
     {
         public int id { get; set; }
-        public int reading { get; set; }
+        public decimal reading { get; set; }
         public System.DateTime reading_date { get; set; }
         public int sensor_id { get; set; }
